fix(gfx): add Window mouse position accessors, unbreaking the app build
Phase 47 Task 10 (hover feature) assumed gfx::Window exposes mouse_x() and
mouse_y() accessors in both image_viewer.cpp and gallery_grid.cpp, but
these methods did not exist. Add them using SDL3's SDL_GetMouseState(),
which returns float* in SDL3 and is window-relative. Remove now-redundant
static_cast<float>() wrappers at both call sites since the accessors
return float directly.

The app was uncompilable; tests remained green only because premake5.lua
enumerates osv_tests files individually, excluding image_viewer.cpp and
gallery_grid.cpp from the test target.

+++ src/ui/gallery_grid.cpp
@@ -1173,7 +1173,7 @@ void GalleryGrid::update(double dt)
 
     // Update hover animation. The gate tracks when to start; playback handles the animation.
     // Phase 47 Task 10: only one hover animation at a time.
-    const int tile = hit_test(static_cast<float>(win_.mouse_x()), static_cast<float>(win_.mouse_y()));
+    const int tile = hit_test(win_.mouse_x(), win_.mouse_y());
     if (hover_gate_.update(tile, dt)) {
         start_hover_animation(tile);
     } else if (hover_gate_.active_tile() != hover_gif_tile_) {
